Move canDeleteMessage module
